backend/requirements.txt
onnxruntime
Pillow
flask
flask_socketio
opencv-python
mediapipe
cvzone
python-dotenv